### What problem does this PR solve?

Issue Number: close #10137

Problem Summary:
before
1. low token notify doesn't work becuase `wait_for` ignore low token signal
2. 5s for computing avg speed is too long, which may not reflect the latest ru speed.
3. acquire token from GAC every 5s

after:
1. `mainLoop` thread will compute ru speed every 1s
2. `requestGACLoop` thread is dedicated to sending gac thread
4. only acquire token from GAC when got low token signal
5. change `min_wait_dura_ms ` from 100ms to 10ms.


### What is changed and how it works?

```commit-message

```

### Check List

Tests <!-- At least one of them must be included. -->

- [ ] Unit test
- [ ] Integration test
- [x] Manual test (add detailed scripts or steps below)
1. create tidb cluster with 2 tiflash(16c)
6. load tpch50g workload
7. run the following scripts: https://github.com/guo-shaoge/tiflash_resource_control_test_tool, detailed test report is internal
    1. [high.change.sh](https://github.com/guo-shaoge/tiflash_resource_control_test_tool/blob/master/high.change.sh)
        
        ![image](https://github.com/user-attachments/assets/706e14fd-9cb8-4e89-9a2b-a633dc63ce63)

    4. [periodically.change.sh](https://github.com/guo-shaoge/tiflash_resource_control_test_tool/blob/master/periodically.change.sh) 
    
        ![image](https://github.com/user-attachments/assets/a29ed8cf-e5d3-4b26-aacd-00335eb988ec)

    6. [priority.change.sh](https://github.com/guo-shaoge/tiflash_resource_control_test_tool/blob/master/priority.change.sh)
        ![image](https://github.com/user-attachments/assets/4911eba7-db9f-46bb-b898-057c70d3d9f3)

    8. [burstable.change.sh](https://github.com/guo-shaoge/tiflash_resource_control_test_tool/blob/master/burstable.change.sh)
        ![image](https://github.com/user-attachments/assets/bcb6eae9-1013-4644-8cd2-f71fdec54a98)


- [ ] No code

Side effects

- [ ] Performance regression: Consumes more CPU
- [ ] Performance regression: Consumes more Memory
- [ ] Breaking backward compatibility

Documentation

- [ ] Affects user behaviors
- [ ] Contains syntax changes
- [ ] Contains variable changes
- [ ] Contains experimental features
- [ ] Changes MySQL compatibility

### Release note

<!-- bugfix or new feature needs a release note -->

```release-note
fix tiflash resource control low token signal miss
```